Moved planet editor script to correct folder
