+++ Assets/Editor/PlanetEditor.cs
@@ -1,0 +1,64 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEditor;
+using UnityEngine;
+
+[CustomEditor(typeof(TempPlanet))]
+public class PlanetEditor : Editor
+{
+    TempPlanet planet;
+    Editor shapeSettingsEditor;
+    Editor colourSettingsEditor;
+
+    public override void OnInspectorGUI()
+    {
+        using (var check = new EditorGUI.ChangeCheckScope())
+        {
+            base.OnInspectorGUI();
+            if (check.changed)
+            {
+                planet.GeneratePlanet();
+            }
+        }
+
+        if (GUILayout.Button("Generate Planet"))
+        {
+            planet.GeneratePlanet();
+        }
+
+        DrawSettingsEditor(planet.shapeSettings, planet.OnShapeSettingsUpdated, ref planet.shapeSettings.foldOut, ref shapeSettingsEditor);
+        DrawSettingsEditor(planet.colourSettings, planet.OnColourSettingsUpdated, ref planet.colourSettings.foldOut, ref colourSettingsEditor);
+    }
+
+    void DrawSettingsEditor(Object settings, System.Action onSettingsUpdated, ref bool foldOut, ref Editor editor)
+    {
+        if (settings != null)
+        {
+            foldOut = EditorGUILayout.InspectorTitlebar(foldOut, settings);
+
+            //Check if anything has changed
+            using (var check = new EditorGUI.ChangeCheckScope())
+            {
+                if (foldOut)
+                {
+                    CreateCachedEditor(settings, null, ref editor);
+                    editor.OnInspectorGUI();
+
+                    if (check.changed)
+                    {
+                        if (onSettingsUpdated != null)
+                        {
+                            onSettingsUpdated();
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    void OnEnable()
+    {
+        planet = (TempPlanet)target;
+    }
+
+}
