+++ Assets/Editor/PlanetEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c7ac413d5f0533d46a0f7f6d862a30aa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
